[ELY-713] Split out common message parsing code so it can be re-used for the HTTP Digest mechanism.
Also cleaned up import ordering.
